+++ twitter/routes/web.php
@@ -38,11 +38,11 @@ Route::group(['prefix' => 'tweet', 'middleware' => 'auth'], function (){
     //ツイート投稿
     Route::post('', [App\Http\Controllers\TweetController::class, 'create'])->name('tweets.create');
     //ツイート一覧表示
-    Route::get('/show', [App\Http\Controllers\TopController::class, 'index'])->name('tweets.show');
+    Route::get('/show', [App\Http\Controllers\TweetController::class, 'index'])->name('tweets.show');
     //ツイート詳細画面に遷移
     Route::get('/{id}/details', [App\Http\Controllers\TweetController::class, 'findByTweetId'])->name('tweet.details');
     //ツイートを更新して、ツイート一覧表示に遷移
-    Route::put('/{id}/update', [App\Http\Controllers\TopController::class, 'update'])->name('tweet.update');
+    Route::put('/{id}/update', [App\Http\Controllers\TweetController::class, 'update'])->name('tweet.update');
 });
